devicetree: cpu nodes may have 64-bit regs
Reported-by: Andre Przywara <andre.przywara@arm.com>
Cc: Laurent Vivier <lvivier@redhat.com>
Cc: Thomas Huth <thuth@redhat.com>

+++ lib/arm/setup.c
@@ -40,7 +40,7 @@ int mpidr_to_cpu(uint64_t mpidr)
 	return -1;
 }
 
-static void cpu_set(int fdtnode __unused, u32 regval, void *info __unused)
+static void cpu_set(int fdtnode __unused, u64 regval, void *info __unused)
 {
 	int cpu = nr_cpus++;
